# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "TODO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/company/{company_id}/task/create": {
            "post": {
                "description": "タスクを作成する。編集者のみ可能。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスクの作成",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "企業ID",
                        "name": "company_id",
                        "in": "path"
                    },
                    {
                        "description": "タスク作成用リクエスト",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.TaskCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/{company_id}/task/list": {
            "get": {
                "description": "閲覧可能なタスクの一覧を企業IDから取得する。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "企業のタスク一覧の取得",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "企業ID",
                        "name": "company_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/{company_id}/task/list_by_assigned_user_id/{assigned_user_id}": {
            "get": {
                "description": "閲覧可能なタスクの一覧を割り当てユーザIDから取得する。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "ユーザに割り当てられたタスク一覧の取得",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "企業ID",
                        "name": "company_id",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "ユーザID",
                        "name": "assigned_user_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/{company_id}/task/{task_id}": {
            "get": {
                "description": "閲覧可能なタスクの情報をIDから取得する。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスクの取得",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "企業ID",
                        "name": "company_id",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "タスクID",
                        "name": "task_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/{company_id}/task/{task_id}/status/{task_status}": {
            "put": {
                "description": "タスクのステータスを指定した状態へと更新する。編集者のみ可能。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスクステータスの更新",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "企業ID",
                        "name": "company_id",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "タスクID",
                        "name": "task_id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "タスクステータス",
                        "name": "task_status",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company/{company_id}/task/{task_id}/update": {
            "put": {
                "description": "タスクの情報を更新する。編集者のみ可能。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスクの更新",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "企業ID",
                        "name": "company_id",
                        "in": "path"
                    },
                    {
                        "type": "integer",
                        "description": "タスクID",
                        "name": "task_id",
                        "in": "path"
                    },
                    {
                        "description": "タスク更新用リクエスト",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.TaskUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.TaskCreate": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "limit_date": {
                    "type": "string"
                },
                "person_in_charge_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "request.TaskUpdate": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "limit_date": {
                    "type": "string"
                },
                "person_in_charge_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
        "todo_api_internal_adapter_inbound_http_model.Company": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "todo_api_internal_adapter_inbound_http_model.Task": {
            "type": "object",
            "properties": {
                "create_at": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.User"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "limit_date": {
                    "type": "string"
                },
                "person_in_charge": {
                    "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.User"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "updator": {
                    "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.User"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
        "todo_api_internal_adapter_inbound_http_model.User": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/todo_api_internal_adapter_inbound_http_model.Company"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        }
    }
}